@@ -17,12 +17,12 @@
   },
   "devDependencies": {
     "@types/node": "^16.7.6",
-    "copy-webpack-plugin": "^9.0.1",
+    "copy-webpack-plugin": "11.0.0",
     "css-loader": "^6.2.0",
     "style-loader": "^3.2.1",
-    "webpack-dev-server": "^4.1.0",
+    "webpack-dev-server": "^4.5.1",
     "ts-node": "^10.2.1",
     "typescript": "^4.4.2",
-    "webpack": "^5.51.1"
+    "webpack": "^5.89.0"
   }
 }
