@@ -1,12 +1,12 @@
 import { App, OpenApiV3 } from 'cdkoa';
-import { Info, Paths } from '../../cdkoa/lib/openapi-v3';
 import description from '../lib/info/description';
 import pet from '../lib/paths/pet';
 import petFindByStatus from '../lib/paths/pet-findByStatus';
 import petFindByTags from '../lib/paths/pet-findByTags';
 import petPetId from '../lib/paths/pet-petId';
+import {Info, Paths} from "cdkoa/lib/openapi-v3";
 
-var app = new App({});
+const app = new App({});
 
 new OpenApiV3.Document(app, 'my-api', {
   info: new Info({
@@ -16,7 +16,7 @@ new OpenApiV3.Document(app, 'my-api', {
     termsOfService: 'http://swagger.io/terms/'
   }),
   paths: new Paths({
-    '/pet': pet,
+    '/pets': pet,
     '/pet/{petId}': petPetId,
     '/pet/findByStatus': petFindByStatus,
     '/pet/findByTags': petFindByTags,
